Huge graphics and gameplay update
Users can now connect to each other in 1v1 games. They can see the other
user update in real time. There is a puck which they can interact with -
it can be hit by each player, and they can score goals for points, which
each user is notified of. There are now graphics for each user's paddle
and the background, as well as a background for the puck.